Added a confirmation window for deleting surveys. (#169)
This is a quick solution for ensuring people don't accidentally delete
surveys. I ran the playwright tests to ensure that this doesn't affect
adding/removing surveys and it does not. I do not think a separate test
is necessary for this as it is just a OS/browser prompt.

(I don't think any other deletions require confirmation so):

Closes #161

---------

Co-authored-by: Carson McManus <carson.mcmanus1@gmail.com>