Drag & drop (#178)
This is allows for items in a survey being edited to be reordered using
drag and drop functionality. It is fully functional as is, however, the
items only work if they are dragged to another item's textbox and when
they get highlighted is super wonky. It works and updates the questions
array to be saved, but the UI bugs need to be fixed before merging.



----
Update 4/25:
The drag & drop is fully functional with both the backend and the UI.
The only thing not included in this PR is a test to test out drag and
drop functionality, however, while potentially possible in Playwright,
it seems to be a bit buggy and weird. I can give a test a solid go, but
as of right now, I have not been able to break the code manually.

Closes #94
Closes #158